+++ demo/platform/wayland.h
@@ -2,6 +2,7 @@
 
 #include "cursor-shape-v1-client-protocol.h"
 #include "fractional-scale-v1-client-protocol.h"
+#include "linux-dmabuf-v1-client-protocol.h"
 #include "viewporter-client-protocol.h"
 #include "xdg-shell-client-protocol.h"
 #include <linux/input-event-codes.h>